+++ protected/humhub/config/common.php
@@ -24,7 +24,7 @@ if (!defined('PKCS7_DETACHED')) {
 
 $config = [
     'name' => 'HumHub',
-    'version' => '1.14.0',
+    'version' => '1.15.0-dev',
     'minRecommendedPhpVersion' => '7.4',
     'minSupportedPhpVersion' => '7.4',
     'basePath' => dirname(__DIR__) . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR,
@@ -269,6 +269,12 @@ $config = [
         'enablePjax' => true,
         'dailyCronExecutionTime' => '18:00',
     ],
+    'container' => [
+        'definitions' => [
+            //todo: Remove after Yii 2.0.48 release
+            \yii\validators\DateValidator::class => humhub\components\validators\DateValidator::class,
+        ]
+    ]
 ];
 
 return $config;